requirements.txt
Django==1.6.5
pygments==1.6
djangorestframework==2.3.14
